Added support for single input tokenization
